refactor(tuner): shadcn select and checkbox everywhere — no raw form controls
